Separate the collections-edit components #37
